Don't handle Homebrew `bin` directory in ZSH config
On Mac OS X, I'll just handle this using `/etc/paths`*. To be specific,
it should be the first thing in that file, causing it to appear at the
front of my `PATH` since my ZSH config uses the `PATH` builder utility.

It is security-conscious to note that including it first means that
anyone with control over the Homebrew `bin` directory could maliciously
include something called, say, `sudo`, which would "win" when the shell
searches for which `sudo` to run and could therefore steal my password.
But I assume that on my mostly-single-user machines there are no
malicious users.

Furthermore, it is important to include the Homebrew `bin` directory
toward the front of my `PATH` so that it can provide newer and/or better
versions of utilities (such as Git).

[*] I read about the `/etc/paths` trick somewhere on the Web, but I'm
    not sure where.