Improve ZSH shell/window behavior via precmd array
Clean up how precmd behavior is handled by using the precmd array
instead of the singular precmd function.

Also improve/correct window title setting, which did not work for
Ghostty using the escape sequences from before.

Also introducing ZSH precmd configuration to ensure that the cursor is a
blinking vertical bar. This works "out of the box" for some terminal
emulators... but not for tmux within Ghostty. But now it's good!